Seperate x86-64 specific code out of generic impl
