+++ src/arch/x86/desc.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define RAM_BASE 0
