+++ src/vm.h
@@ -18,10 +18,17 @@ typedef struct {
     struct virtio_blk_dev virtio_blk_dev;
 } vm_t;
 
+int vm_arch_init(vm_t *v);
+int vm_arch_cpu_init(vm_t *v);
+int vm_arch_init_platform_device(vm_t *v);
+int vm_arch_load_image(vm_t *v, void *image, size_t size);
+int vm_arch_load_initrd(vm_t *v, void *initrd, size_t size);
+
 int vm_init(vm_t *v);
 int vm_load_image(vm_t *v, const char *image_path);
 int vm_load_initrd(vm_t *v, const char *initrd_path);
 int vm_load_diskimg(vm_t *v, const char *diskimg_file);
+int vm_late_init(vm_t *v);
 int vm_run(vm_t *v);
 int vm_irq_line(vm_t *v, int irq, int level);
 void *vm_guest_to_host(vm_t *v, void *guest);
